requirements.txt
fastapi>=0.104.0
uvicorn>=0.23.2
jinja2>=3.1.2
sqlalchemy>=2.0.23
alembic>=1.12.1
pydantic>=2.4.2
python-jose>=3.3.0
passlib>=1.7.4
python-multipart>=0.0.6
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
aiohttp>=3.8.6
python-telegram-bot>=20.6
pika>=1.3.2
geopy>=2.4.0
matplotlib>=3.8.0
pandas>=2.1.1
email-validator==2.1.0 